Sync up to 0ef3ce173a0addf7d9de7e2ccc80e316789251b7 in v0.8.5.

+++ selfdrive/test/setup_device_ci.sh
@@ -26,11 +26,11 @@ if [ ! -d "$SOURCE_DIR" ]; then
   git clone --depth 1 https://github.com/commaai/openpilot.git "$SOURCE_DIR"
 fi
 
-# clear scons cache dirs that haven't been written to in one day
-#cd /tmp && find -name 'scons_cache_*' -type d -maxdepth 1 -mtime +1 -exec rm -rf '{}' \;
-
-# this can get really big on the CI devices
-rm -rf /data/core
+if [ -f "/EON" ]; then
+  rm -rf /data/core
+  rm -rf /data/neoupdate
+  rm -rf /data/safe_staging
+fi
 
 # set up environment
 cd $SOURCE_DIR
